Threading, any octet change, output to a file
