+++ Backend_Code/h.py
@@ -1,0 +1,5 @@
+def help():
+    f = open('help.txt', 'r')
+    file_contents = f.read()
+    print (file_contents)
+    f.close()
