--- help.txt
@@ -1,14 +1,0 @@
-Syntax:
-    python3 ping.py [{-h}, {--help}, {-si}, {-start_ip}, {-start}, {-ei}, {-end_ip}, {-end}]
-
--h, --help: shows this page
--si, -start_ip, -start: the start ip adress for the range
--ei, -end_ip, -end: the end ip for the range
-
-Important notes:
-    The range will only work for the last octet for the ip
-
-V2 comming next:
-    will range no matter what octet is changed and fix Important notes
-    support ipV6
-    include threading
