+++ requirements.txt
@@ -4,4 +4,4 @@ sys
 cli_args_system
 threading
 time
-
+pyfiglet
